fix type in TreeNode

+++ data_structures/binary_tree/distribute_coins.py
@@ -45,7 +45,7 @@ from dataclasses import dataclass
 
 @dataclass
 class TreeNode:
-    data: float
+    data: int
     left: TreeNode | None = None
     right: TreeNode | None = None
